Refactor: DatePicker에 formattingDate, formattingTime 함수 utils로 빼기
